CategoryDialog: Refactor to use sx prop instead of makeStyles, and to remove redundancy between CreateCategoryDialog and EditCategoryDialog
