--- src/components/annotator/CategoriesList/CreateCategoryListItem/CreateCategoryListItem/index.ts
@@ -1,1 +1,0 @@
-export { CreateCategoryListItem } from "./CreateCategoryListItem";
