+++ src/components/annotator/CategoryDialog/CategoryDialog/index.ts
@@ -1,0 +1,1 @@
+export { CategoryDialog } from "./CategoryDialog";
