+++ src/components/annotator/CategoryDialog/index.ts
@@ -1,0 +1,2 @@
+export { CreateCategoryDialog } from "./CreateCategoryDialog";
+export { EditCategoryDialog } from "./EditCategoryDialog";
